refactor(core): share code between ComponentRef and targeted instructions
This refactor reorganizes code such that the dynamic bindings logic and
targeted listener instruction can share most of the code.
