+++ packages/core/test/bundling/defer/bundle.golden_symbols.json
@@ -468,6 +468,7 @@
   "init_di_attr",
   "init_di_setup",
   "init_directive",
+  "init_directive_outputs",
   "init_directives",
   "init_directives2",
   "init_discovery",
@@ -583,6 +584,7 @@
   "init_linker",
   "init_list_reconciliation",
   "init_listener",
+  "init_listeners",
   "init_local_compilation",
   "init_locale_data_api",
   "init_locale_en",
