+++ packages/core/test/bundling/forms_reactive/bundle.golden_symbols.json
@@ -409,7 +409,6 @@
   "getTStylingRangePrev",
   "getTView",
   "getViewRefs",
-  "handleError",
   "handleStoppedNotification",
   "handleUnhandledError",
   "hasApplyArgsData",
