+++ packages/core/test/bundling/forms_template_driven/bundle.golden_symbols.json
@@ -397,7 +397,6 @@
   "getTStylingRangePrev",
   "getTView",
   "getViewRefs",
-  "handleError",
   "handleStoppedNotification",
   "handleUnhandledError",
   "hasApplyArgsData",
