+++ packages/core/test/bundling/router/bundle.golden_symbols.json
@@ -486,7 +486,6 @@
   "getTView",
   "getTokenOrFunctionIdentity",
   "getViewRefs",
-  "handleError",
   "handleStoppedNotification",
   "handleUnhandledError",
   "hasApplyArgsData",
