+++ packages/core/test/bundling/todo/bundle.golden_symbols.json
@@ -339,7 +339,6 @@
   "getTStylingRangePrev",
   "getTView",
   "getViewRefs",
-  "handleError",
   "handleStoppedNotification",
   "handleUnhandledError",
   "hasApplyArgsData",
